+++ LEDGeometry/Projection.cpp
@@ -12,7 +12,7 @@ void radial_projection(Shape* shape, uint8_t* projection, uint8_t resolution,
     float min_dist = shape->min_distance(center_x, center_y);
     float max_dist = shape->max_distance(center_x, center_y);
     float factor = (resolution - 1) / (max_dist - min_dist);
-    for (uint8_t i = 0; i < shape->n_points(); i++) {
+    for (int i = 0; i < shape->n_points(); i++) {
         float dist = shape->distance(i, center_x, center_y);
         projection[i] = (uint8_t)round((dist - min_dist) * factor);
     }
@@ -31,7 +31,7 @@ void parallel_projection(Shape* shape, uint8_t* projection, uint8_t resolution,
     float min_proj = shape->min_projection(direction_x, direction_y);
     float max_proj = shape->max_projection(direction_x, direction_y);
     float factor = (resolution - 1) / (max_proj - min_proj);
-    for (uint8_t i = 0; i < shape->n_points(); i++) {
+    for (int i = 0; i < shape->n_points(); i++) {
         float proj = shape->projection(i, direction_x, direction_y);
         projection[i] = (uint8_t)round((proj - min_proj) * factor);
     }
